refactor : CarRacingService 클래스 리팩토링  - CarService 클래스 제거 (Cars 클래스가 생성자를 통해 기존 역할을 대신할 수 있음)  - CarRacingService 의 비즈니스 로직을 RacingResult 클래스의 생성자 내부로 옮김    (자연스럽게 하위 데이터 객체들의 빌드처리가 한 클래스 내부에서 모두 일어나게 됨)
