+++ src/main/java/carracing/constants/CarRacingConstant.java
@@ -3,5 +3,4 @@ package carracing.constants;
 public class CarRacingConstant {
     public static final int MIN_CAR_REGISTER_COUNT = 1;
     public static final int MIN_ROUND_COUNT = 1;
-    public static final int ROUND_START_NUMBER = 1;
 }
