+++ src/main/java/carracing/service/CarRacingService.java
@@ -1,10 +1,8 @@
 package carracing.service;
 
-import carracing.constants.CarRacingConstant;
-import carracing.domain.Car;
-import carracing.domain.CarService;
 import carracing.domain.Cars;
-import carracing.service.dto.*;
+import carracing.service.dto.RacingRegisterInfo;
+import carracing.service.dto.RacingResult;
 
 /**
  * CarRacingService
@@ -16,17 +14,15 @@ import carracing.service.dto.*;
  */
 public class CarRacingService {
 
-    private CarService         carService;
     private CarRacingValidator carRacingValidator;
 
     public CarRacingService() {
-        this.carService = new CarService();
         this.carRacingValidator = new CarRacingValidator();
     }
 
     public RacingResult executeCarRacing(RacingRegisterInfo racingRegisterInfo) {
         validate(racingRegisterInfo);
-        return getRacingResult(racingRegisterInfo.getRoundCount(), carService.registerCars(racingRegisterInfo.getRacingCarCount()));
+        return new RacingResult(racingRegisterInfo.getRoundCount(), new Cars(racingRegisterInfo.getRacingCarCount()));
     }
 
     private void validate(RacingRegisterInfo racingRegisterInfo) {
@@ -34,24 +30,4 @@ public class CarRacingService {
         carRacingValidator.validateRoundCount(racingRegisterInfo.getRoundCount());
     }
 
-    private RacingResult getRacingResult(int roundCount, Cars cars) {
-        RacingResult racingResult = new RacingResult();
-        for (int i = CarRacingConstant.ROUND_START_NUMBER; i <= roundCount; i++) {
-            racingResult.registerRoundResult(executeRound(i, cars));
-        }
-        return racingResult;
-    }
-
-    private RoundResult executeRound(int roundNumber, Cars cars) {
-        return new RoundResult(roundNumber, executeScoring(cars));
-    }
-
-    private RacingScores executeScoring(Cars cars) {
-        RacingScores racingScores = new RacingScores();
-        for (Car car : cars.getCarList()) {
-            car.drive();
-            racingScores.registerRacingScore(car);
-        }
-        return racingScores;
-    }
 }
